requirements.txt
asgiref==3.7.2
docx2pdf==0.1.8
Django==4.2.7
djangorestframework==3.14.0
python-docx==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
